extend RewindReader with ReaderAt, Seeker methods
Allows blob storage backends to optimise with concurrent chunking.  This
message will focus on the b2 backend, though other blob storage backends
can likely be made to work in a similar fashion.

The b2 backend degrades to a set of large in-process buffers without an
io.ReadSeeker.  Large portions of the file are read into process memory
in order to allow concurrent uploaders to work on different segments at
once.  This is a needless waste of memory.  The kernel buffer cache
should be all that we need.

Exposing io.ReaderAt is an optimisation that further avoids a needless
layer of mutex contention in the b2 backend.