b2: enable concurrent chunk uploads within a single Save
Improves aggregate throughput on machines with few logical CPU cores.

Reuse the existing b2.connections knob in interests of simplicity.  This
reuse is not technically correct, for the total number of connections to
b2 will now exceed b2.connections on machines with multiple logical CPU
cores.  Total b2 connection count will equal the number of blob savers
multiplied by the value of b2.connections.

This is probably fine unless you are on a monster with >32 cores.